python/requirements.txt
boto3==1.9.0
botocore==1.12.62
Pillow==6.2.2
